+++ CHANGES.md
@@ -5,6 +5,7 @@
 
 ## 0.0.50
 
+-   Add readiness and liveness probes to all services
 -   Added configurable argument to `magda-web-server` module to accept Google Analytics IDs
 -   Make datasetSearchSuggestionScoreThreshold and searchResultsPerPage runtime configurable
 -   Hide dataset source for csv datasets